+++ src/Main.js
@@ -12,16 +12,24 @@ define(function (require, exports, module) {
     // Brackets Modules
     var EditorManager = brackets.getModule("editor/EditorManager");
     var ExtensionUtils = brackets.getModule("utils/ExtensionUtils")
+    var CommandManager = brackets.getModule("command/CommandManager")
 
     // Local modules
     var Channel = require("src/utils/Channel")
     var Events = require("src/utils/Events")
     var Utils = require("src/utils/Utils")
+    var C = require("src/utils/Constants")
+    var S = require("strings")
     require('src/Zotero')
-    require('src/Panel')
+    require('src/ui/Panel')
+    require('src/ui/SearchPanelView')
+    require('src/ui/CitePanelView')
     require('src/Document')
     require('src/Bibliography')
-    require('src/SettingsDialog')
+    require('src/ui/SettingsDialog')
+
+    // private variables
+    var $icon = null;
 
     /**
      * Channeling `activeEditorChange` Brackets' Event to the
@@ -31,11 +39,21 @@ define(function (require, exports, module) {
         Channel.Extension.trigger(Events.EVENT_EDITOR_CHANGE, newEd)
     })
 
+    function _handleZoteroIconClassToggle() {
+        $icon.toggleClass('active')
+    }
+
     function _init() {
         // registering all commands to use Extension-wide Radio Channel
         Utils.registerCommandsAndKeyBindings()
         // loading octicons css
-        ExtensionUtils.loadStyleSheet(module, '../styles/icons/octicons/octicons.css');
+        ExtensionUtils.loadStyleSheet(module, '../styles/main.less');
+        // Displaying icon in the Project Panel and attach `click` handler
+        $icon = $('<a id="zotero-toolbar-icon" href="#" title="' + S.UI_TOOLTIP + '"></a>')
+                .appendTo($("#main-toolbar .buttons"))
+                .on("click", function () { CommandManager.execute(C.CMD_ID_TOGGLE_PANEL) })
+        // complying to toggle class on icon
+        Channel.Extension.comply( Events.COMMAND_TOGGLE_ICON_STATE, _handleZoteroIconClassToggle )
         // wake up everyone
         Channel.Extension.trigger(Events.EVENT_INIT)
     }